Fixed a compile warning

+++ ui/main-window.cpp
@@ -78,7 +78,7 @@ void MainWindow::initWidgets()
     this->levelSpeed = new LevelComboBox(this);
     this->goButton = new QPushButton("Randomize!", this);
     this->spriteSelect = new QComboBox(this);
-    for (int i=0; i < sprite_count(); ++i) {
+    for (size_t i=0; i < sprite_count(); ++i) {
        spriteSelect->addItem(dwr_sprite_names[i]);
     }
     this->tabWidget = new QTabWidget(this);
